[useEvent] Non-stable function identity
Since useEvent shouldn't go in the dependency list of whatever is
consuming it (which is enforced by the fact that useEvent functions are
always locally created and never passed by reference), its identity
doesn't matter. This change updates our implementation to not store the
wrapped event function in memoized state, rather the wrapper is
recreated on every update. Effectively, this PR is a runtime assertion
that you can't rely on the return value of useEvent to be stable.